After further consideration I decided to set the default value for language in the app manager to be None. This is because if we set it to a default value (en) it can create confusion when the user passes in a locale value, and now the app manager has both a language and a locale.
Functionally it does not matter since duckling sets the preference to locale, but to avoid confusion I have added extra logic to set the default language only if the locale is not set.

I have also added tests to make sure that we can override the app manager's language/locale using params.