Reprocessing queries for slot-filling (#156)
This PR allows slot-filling to use allowed_intents and reprocess a query when the retry logic is exhausted. 

This comes from the observation that the in-built reprocess does not work with allowed_intents set from the previous turn and always reprocesses to those specified intents. This fix by-passes the reprocess call and directly calls the apply_handler functionality in the DM with a repeat nlp.process call and creation of a new request object that is passed to the DM as a new query turn.

Also, fixes a validation bug and tests for all of the above.