AutoFill Bugfix - Add a CallableRegistry class to MindMeld (#353)
Fixes issue #352 .

The above bug was caused due to serialization of the custom eval function as a string while the automatic slot filling functionality was calling functions directly. This solution creates a registry class to register the custom functions at app initialization. It then refers to this mapping at runtime when utilizing the slot filling functionality to call any relevant functions.

Modifications:
Added a Registry class to core.py for registration needs for MindMeld. Currently, it has a function register method that the custom_eval functionality is using to register custom functions from developers (custom functions in the DM can be registered using the @app.register_func() decorator.

For future changes, we can use this class to replace the register_annotator/augmentor functionalities.

* updated FormEntity schema and callable check in core.py

* serialize custom_eval

* add custom func registration support

* update registry docstring

* update docs with registration custom eval

* added getter/setter for CallableRegistry class