Remove SER reference from EntityLabelEncoder.
This bug was introduced in Mindmeld 4.3.0 with the SER refactor, then
further modified in a change to make models built against Mindmeld < 4.3
compatible by assigning a `system_entity_recognizer` attribute.

This removes the `system_entity_recognizer` attribute from
`EntityLabelEncoder`. The presence of the attribute causes the referent
to be serialized at with the classifer at the time it is saved. Because
the referent SER also includes an attritube for the URL it should use,
the result is that the classifier will use its own copy of the SER
pointing to the URL used at save time, and cannot be updated to use a
different URL.

This is reproducible with the following procedure, assuming you have a
working Mindmeld application that requires SER inference for some
intents.

* In the application's `config.py`, add a custom
  `system_entity_recognizer` in `NLP_CONFIG` to override the default
  (see example below) using an invalid URL.
* Build models for that application.
* Remove the custom `system_entity_recognizer` from `NLP_CONFIG`.
* Run the application, then make a request that requires inference via
  the SER.

`config.py` snippet:

```python
NLP_CONFIG = {
    'system_entity_recognizer': {
        'type': 'duckling',
        'url': os.environ.get('DUCKLING_URL', 'http://this.url.does.not.work/parse'),
    },
}
```

The request should fail with an error in the application:

`Unable to connect to the system entity recognizer. Make sure it's running by typing 'mindmeld num-parse' at the command line.`

For intents that don't have entity classifiers (or don't rely on SER),
inference should work normally without errors.

A notable caveat for this change--I suspect this will cause
compatibility issues for models built prior to the release because the
`system_entity_recognizer` attribute no longer exists on
`EntityLabelEncoder`, yet Pickle will expect to deserialize that
attribute. If we incorporate this (which I think we must), it should
come with clear documentation explaining that models in the new version
are no longer compatible and need to be rebuilt.

SER refactor in 4.3.0: https://github.com/cisco/mindmeld/pull/88
Compatibility change in 4.3.2: https://github.com/cisco/mindmeld/pull/209